Minor cleanup / Fixes
Fixed an issue where B&S would not load due to the original dll still being around trying to load in.

Fixed an issue where hidden files would cause an error, it will not unhide the file itself if not already.

Changed some foreach loops to for for better performance.

ORIGINAL-XX.dll will no longer be around, no need for it.